Split web API contract parsers
